+++ src/js/chuck-norris/apis.js
@@ -2,8 +2,8 @@
 import { type AjaxResponse } from 'rxjs';
 import { ajax } from 'rxjs/observable/dom/ajax';
 
-export const RANDOM_JOKE_SERVER = 'https://api.icndb.com';
-export const RANDOM_JOKE_URI = '/jokes/random';
+export const RANDOM_JOKE_SERVER: string = 'https://api.icndb.com';
+export const RANDOM_JOKE_URI: string = '/jokes/random';
 
 // One way to decode HTML
 // See http://stackoverflow.com/questions/7394748/whats-the-right-way-to-decode-a-string-that-has-special-html-entities-in-it