+++ src/js/chuck-norris/reducer.js
@@ -1,14 +1,13 @@
 // @flow
 import createReducer from '../app/utils/create-reducer';
-import ChuckNorrisRecord from './models/chuck-norris-record';
-import { ACTION_TYPES, type GetJokeAction } from './actions';
+import makeChuckNorrisRecord, { type ChuckNorrisRecord } from './models/chuck-norris-record';
+import ACTION_TYPES, { type GetJokeAction } from './types';
 
-const handleAction = (
-  state: ChuckNorrisRecord,
-  action: GetJokeAction,
-): ChuckNorrisRecord => state.merge(action.state);
+type HandleActionFn = (state: ChuckNorrisRecord, action: GetJokeAction) => ChuckNorrisRecord;
 
-export default createReducer(new ChuckNorrisRecord(), {
+const handleAction: HandleActionFn = (state, action) => state.merge(action.state);
+
+export default createReducer(makeChuckNorrisRecord(), {
   [ACTION_TYPES.GET_JOKE]: handleAction,
   [ACTION_TYPES.GET_JOKE_SUCCEEDED]: handleAction,
   [ACTION_TYPES.GET_JOKE_FAILED]: handleAction,