+++ src/js/error/components/PageNotFoundError.js
@@ -3,7 +3,7 @@ import React from 'react';
 import Typography from 'material-ui/Typography';
 
 const PageNotFoundError = () => (
-  <Typography type="display2" gutterBottom>
+  <Typography variant="display2" gutterBottom>
     The page you are looking for does not exist.
   </Typography>
 );