+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.11.0 - 2017-XX-XX
 
 * When running `yarn ci`, don't run Flow because it has higher chances of breaking when attempting to run Flow server on CI servers such as Jenkins and TFS Build.
+* Webpack - Applied tree-shaking to `vendor.js` to significantly reduce file size from 1.73MB to 768KB.
 
 ## 0.10.0 - 2017-11-29