+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * When running `yarn ci`, don't run Flow because it has higher chances of breaking when attempting to run Flow server on CI servers such as Jenkins and TFS Build.
 * Webpack - Applied tree-shaking to `vendor.js` to significantly reduce file size from 1.73MB to 768KB.
-* Configured ESLint rules to allow quotes around number... for example: `const a = { '1919': 'value' };`
+* Webpack - Replaced `Happypack` with `HardSourceWebpackPlugin` to further improve build time from 40s to 5s.
+* ESLint - Configured rules to allow quotes around number... for example: `const a = { '1919': 'value' };`
 * Mocha - Replaced deprecated `--compilers` with `--require`. See https://github.com/mochajs/mocha/wiki/compilers-deprecation
 
 ## 0.10.0 - 2017-11-29